kgsl: add conservative governor [squashed]
This is a combination of 3 commits.

The 1st commit's message is:
kgsl: Implement conservative scaling policy
This behaves similarly to the conservative cpufreq governor,
which is similar to how most Mali GPU scaling algorithms behave.
The "trustzone" policy that is default relies on an undocumented
black box to make scaling decisions.
How to enable:
echo conservative > /sys/devices/platform/kgsl-3d0.0/kgsl/kgsl-3d0/pwrscale/policy
Conservative attr group:
/sys/devices/platform/kgsl-3d0.0/kgsl/kgsl-3d0/pwrscale/policy_config/conservative
Polling interval (10-1000):
/sys/devices/platform/kgsl-3d0.0/kgsl/kgsl-3d0/pwrscale/policy_config/conservative/polling_interval
Enable stats printing (causes microlags on low polling intervals):
echo 1 > /sys/devices/platform/kgsl-3d0.0/kgsl/kgsl-3d0/pwrscale/policy_config/conservative/print_stats
NOTE:
- perms are set to 666 for testing purposes

Change-Id: I9d854a73981515fc9850104302804928b04c4687
Signed-off-by: Andrew Dodd <atd7@cornell.edu>

This is the 2nd commit message:
kgsl: conservative: change perms to 664

Change-Id: Ie023a6b0f3bf51c4c59030a79ae4ebbbce08201c

This is the 3rd commit message:
kgsl_pwrscale: Rework policy sysfs implementation
In the current implementation, the pwrscale policy kobj is
created when the pwrscale driver initializes.  It is then added
to the kobj hierarchy when a governor adds sysfs files.
When a governor removes sysfs files, it frees the kobj.
Then, when a new governor starts up, it attempts to add an
uninitialized kobj.  This fails.
Instead, use a constant kobj for the policy directory,
and add governor-specific attributes under it.  If a
sysfs group is given a name, all attributes in the group
will appear in a subdirectory with that name (see cpufreq
governor tuning variables).  Use this approach for our
governor tuning.
TODO (possibly in a future patch) - Rework this so the
governor attributes can be groups attached directly to
the pwrscale kobj.

Change-Id: I7b8ae6d5e690ee107be037ccaf2e737e52062feb